resumable solving

+++ src/lib/solving.sh
@@ -9,19 +9,21 @@ SAT_HERITAGE_INPUT_KEY=sat_heritage # the name of the input key to access SAT he
 # solves a file
 # measures the solve time
 solve-file(file, solver_name, solver, data_fields=, data_extractor=, timeout=0, ignore_exit_code=, attempts=, attempt_grouper=) {
+    local input output_log timeout_file csv_line timeouts fail_fast
     attempt_grouper=${attempt_grouper:-$(lambda file echo default)}
-    local input
     input="$(input-directory)/$file"
+    output_log=$(mktemp)
+    csv_line="$file,"
+    log "$solver_name: $file"
+    if [[ -f $(output-csv) ]] && grep -qP "^\Q$csv_line\E" "$(output-csv)"; then
+        log "" "$(echo-skip)"
+        return
+    fi
+    log "" "$(echo-progress solve)"
     compile-lambda solver "$solver"
     compile-lambda attempt-grouper "$attempt_grouper"
-    local output_log
-    output_log=$(mktemp)
-    local timeout_file
     timeout_file=$(output-file "$(attempt-grouper "$file").timeout")
-    log "$solver_name: $file" "$(echo-progress solve)"
-    local timeouts
     timeouts="$(wc -l 2>/dev/null < "$timeout_file" || echo 0)"
-    local fail_fast=
     if [[ -n $attempts ]] && [[ $timeouts -ge $attempts ]]; then
         fail_fast=y
     fi
@@ -37,8 +39,7 @@ solve-file(file, solver_name, solver, data_fields=, data_extractor=, timeout=0, 
     if grep -q "^measure_timeout=y" < "$output_log"; then
         echo "$file" >> "$timeout_file"
     fi
-    local csv_line=""
-    csv_line+="$file,$solver_name,$(grep -oP "^measure_time=\K.*" < "$output_log" || echo)"
+    csv_line+="$solver_name,$(grep -oP "^measure_time=\K.*" < "$output_log" || echo)"
     if [[ -n $data_extractor ]]; then
         if [[ -z $fail_fast ]] && { [[ -n $ignore_exit_code ]] || [[ $(grep -oP "^measure_exit_code=\K.*" < "$output_log") -eq 0 ]]; }; then
             compile-lambda data-extractor "$data_extractor"
@@ -50,23 +51,32 @@ solve-file(file, solver_name, solver, data_fields=, data_extractor=, timeout=0, 
         fi
     fi
     rm-safe "$output_log"
-    # technically, this write is unsafe when using parallel jobs.
-    # however, as long as the line is not too long, the write buffer saves us.
+    # technically, this write is unsafe when using parallel jobs
+    # however, as long as the line is not too long, the write buffer saves us
     # see https://unix.stackexchange.com/q/42544/
     echo "$csv_line" >> "$(output-csv)"
 }
 
 # solves a list of files
 solve-files(csv_file, input_extension, solver_name, solver, data_fields=, data_extractor=, timeout=0, jobs=1, ignore_exit_code=, attempts=, attempt_grouper=) {
-    echo -n "${input_extension}_file,${input_extension}_solver,${input_extension}_solver_time" > "$(output-csv)"
-    if [[ -n $data_fields ]]; then
-        echo ",${data_fields//-/_}" >> "$(output-csv)"
-    else
-        echo >> "$(output-csv)"
+    if [[ ! -f $(output-csv) ]]; then
+        echo -n "${input_extension}_file,${input_extension}_solver,${input_extension}_solver_time" > "$(output-csv)"
+        if [[ -n $data_fields ]]; then
+            echo ",${data_fields//-/_}" >> "$(output-csv)"
+        else
+            echo >> "$(output-csv)"
+        fi
     fi
-    table-field "$csv_file" "${input_extension}_file" | grep -v NA$ | sort -V \
-        | parallel -q ${jobs:+"-j$jobs"} "$SRC_DIRECTORY/main.sh" \
-        solve-file "{}" "$solver_name" "$solver" "$data_fields" "$data_extractor" "$timeout" "$ignore_exit_code" "$attempts" "$attempt_grouper"
+    # to avoid the constant overhead from parallelization due to reloading torte, run sequentially if only one job is requested
+    if [[ $jobs -eq 1 ]]; then
+        while IFS= read -r file; do
+            solve-file "$file" "$solver_name" "$solver" "$data_fields" "$data_extractor" "$timeout" "$ignore_exit_code" "$attempts" "$attempt_grouper"
+        done < <(table-field "$csv_file" "${input_extension}_file" | grep -v NA$ | sort -V)
+    else
+        table-field "$csv_file" "${input_extension}_file" | grep -v NA$ | sort -V \
+            | parallel -q ${jobs:+"-j$jobs"} "$SRC_DIRECTORY/main.sh" \
+            solve-file "{}" "$solver_name" "$solver" "$data_fields" "$data_extractor" "$timeout" "$ignore_exit_code" "$attempts" "$attempt_grouper"
+    fi
 }
 
 # runs a solver on a file
